tests/wpt/tests/css/css-conditional/container-queries/top-layer-dialog-container.html

<!doctype html>
<title>CSS Container Queries Test: Top layer element as a @container</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #parent { width: 100px; }
  #dialog {
    container-type: inline-size;
    width: auto;
    border: none;
  }
  #child { color: red; }
  @container (min-width: 200px) {
    #child { color: green; }
  }
</style>
<div id="parent">
  <dialog id="dialog"><span id="child"></span></dialog>
</div>
<script>
  setup(() => assert_implements_size_container_queries());

  test(() => {
    assert_equals(getComputedStyle(child).color, "rgb(255, 0, 0)");
  }, "#dialog initially sized by #containing-block");

  test(() => {
    dialog.showModal();
    assert_equals(getComputedStyle(child).color, "rgb(0, 128, 0)");
  }, "#dialog sized by viewport");
</script>
