LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/top-layer-dialog-container-expected.txt


PASS #dialog initially sized by #containing-block
FAIL #dialog sized by viewport assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
